Creates scan_i18n.js, adds to yarn and husky
scan_i18n.js will iterate over the source files in a given directory to
find any references to the i18n call "t('KEY')" and make a list of
unique keys from the source. It then checks all language files against
that, and notes any misses from the list of in use keys at the end of
its run.