Create knowledgebase page
Attempt to resolve Notion being ill suited and Airtable being very excel
in its public facing UI. Connection requires an API key which has to be
exposed to the user (so we need to create a read-only at some point).
Adds react-icons for kb card icons, airtable for airtable connection.