Update react-link-preview version, set fallback image.
Set max width to prevent weirdness and a fallback to VStack the HStacks
if the text would be too wide to render inside the allotted space of the
card. Out of band, requested another revision of the react-link-preview
to handle non-falsey but non-usable locations that will allow for use of
JSX fallback (so text and the image).