Add OpenSea link to footer (#238)
* Update OpenSea Icon with official svg

Grabbed an updated white SVG logo from:
https://docs.opensea.io/docs/logos

in order to match the styling of the other social
logos on the home page

* Create NewFooter component

This component is refactored from the current index page.

There is another "Footer" component that is not actually displayed in production.

In a future PR we can clean up the code by deleting the old Footer and renaming NewFooter to just "Footer".

* Refactor index page to use NewFooter

This is should be a no-op change. It's a simple
refactor to extract the Footer logic so that the NewFooter component
will be easier to write tests for later.

* Add OpenSea Icon to the footer

This commit resolves https://github.com/Developer-DAO/developerdao.com/issues/237

* Update footer component tests

Rename old tests and add new tests for the NewFooter component.